Updated Assignment 1
Changed custom-named python script to the 01 file even though the files aren't based on each other.
